+++ packages/cloud_firestore/cloud_firestore/ios/Classes/Public/FirestoreMessages.g.h
@@ -110,6 +110,18 @@ typedef NS_ENUM(NSUInteger, PigeonTransactionType) {
 - (instancetype)initWithValue:(PigeonTransactionType)value;
 @end
 
+typedef NS_ENUM(NSUInteger, AggregateType) {
+  AggregateTypeCount = 0,
+  AggregateTypeSum = 1,
+  AggregateTypeAverage = 2,
+};
+
+/// Wrapper for AggregateType to allow for nullability.
+@interface AggregateTypeBox : NSObject
+@property(nonatomic, assign) AggregateType value;
+- (instancetype)initWithValue:(AggregateType)value;
+@end
+
 @class PigeonFirebaseSettings;
 @class FirestorePigeonFirebaseApp;
 @class PigeonSnapshotMetadata;
@@ -121,6 +133,8 @@ typedef NS_ENUM(NSUInteger, PigeonTransactionType) {
 @class PigeonTransactionCommand;
 @class DocumentReferenceRequest;
 @class PigeonQueryParameters;
+@class AggregateQuery;
+@class AggregateQueryResponse;
 
 @interface PigeonFirebaseSettings : NSObject
 /// `init` unavailable to enforce nonnull fields, see the `make` class method.
@@ -257,6 +271,25 @@ typedef NS_ENUM(NSUInteger, PigeonTransactionType) {
 @property(nonatomic, strong, nullable) NSDictionary<NSString *, id> *filters;
 @end
 
+@interface AggregateQuery : NSObject
+/// `init` unavailable to enforce nonnull fields, see the `make` class method.
+- (instancetype)init NS_UNAVAILABLE;
++ (instancetype)makeWithType:(AggregateType)type field:(nullable NSString *)field;
+@property(nonatomic, assign) AggregateType type;
+@property(nonatomic, copy, nullable) NSString *field;
+@end
+
+@interface AggregateQueryResponse : NSObject
+/// `init` unavailable to enforce nonnull fields, see the `make` class method.
+- (instancetype)init NS_UNAVAILABLE;
++ (instancetype)makeWithType:(AggregateType)type
+                       field:(nullable NSString *)field
+                       value:(NSNumber *)value;
+@property(nonatomic, assign) AggregateType type;
+@property(nonatomic, copy, nullable) NSString *field;
+@property(nonatomic, strong) NSNumber *value;
+@end
+
 /// The codec used by FirebaseFirestoreHostApi.
 NSObject<FlutterMessageCodec> *FirebaseFirestoreHostApiGetCodec(void);
 
@@ -318,12 +351,14 @@ NSObject<FlutterMessageCodec> *FirebaseFirestoreHostApiGetCodec(void);
            parameters:(PigeonQueryParameters *)parameters
               options:(PigeonGetOptions *)options
            completion:(void (^)(PigeonQuerySnapshot *_Nullable, FlutterError *_Nullable))completion;
-- (void)aggregateQueryCountApp:(FirestorePigeonFirebaseApp *)app
-                          path:(NSString *)path
-                    parameters:(PigeonQueryParameters *)parameters
-                        source:(AggregateSource)source
-             isCollectionGroup:(NSNumber *)isCollectionGroup
-                    completion:(void (^)(NSNumber *_Nullable, FlutterError *_Nullable))completion;
+- (void)aggregateQueryApp:(FirestorePigeonFirebaseApp *)app
+                     path:(NSString *)path
+               parameters:(PigeonQueryParameters *)parameters
+                   source:(AggregateSource)source
+                  queries:(NSArray<AggregateQuery *> *)queries
+        isCollectionGroup:(NSNumber *)isCollectionGroup
+               completion:(void (^)(NSArray<AggregateQueryResponse *> *_Nullable,
+                                    FlutterError *_Nullable))completion;
 - (void)writeBatchCommitApp:(FirestorePigeonFirebaseApp *)app
                      writes:(NSArray<PigeonTransactionCommand *> *)writes
                  completion:(void (^)(FlutterError *_Nullable))completion;