+++ packages/cloud_firestore/cloud_firestore/lib/src/aggregate_query.dart
@@ -23,4 +23,10 @@ class AggregateQuery {
   }) async {
     return AggregateQuerySnapshot._(await _delegate.get(source: source), query);
   }
+
+  /// Represents an [AggregateQuery] over the data at a particular location for retrieving metadata
+  /// without retrieving the actual documents.
+  AggregateQuery count() {
+    return AggregateQuery._(_delegate.count(), query);
+  }
 }