+++ packages/cloud_firestore/cloud_firestore_platform_interface/lib/src/method_channel/method_channel_query.dart
@@ -8,6 +8,8 @@ import 'dart:async';
 import 'package:_flutterfire_internals/_flutterfire_internals.dart';
 import 'package:cloud_firestore_platform_interface/cloud_firestore_platform_interface.dart';
 import 'package:cloud_firestore_platform_interface/src/internal/pointer.dart';
+import 'package:cloud_firestore_platform_interface/src/platform_interface/platform_interface_query.dart'
+    as query;
 import 'package:collection/collection.dart';
 import 'package:flutter/services.dart';
 
@@ -272,6 +274,143 @@ class MethodChannelQuery extends QueryPlatform {
       _pigeonParameters,
       _pointer.path,
       pigeonApp,
+      [
+        AggregateQuery(
+          type: AggregateType.count,
+        )
+      ],
+      isCollectionGroupQuery,
+    );
+  }
+
+  @override
+  AggregateQueryPlatform aggregate(
+    AggregateField aggregateField1, [
+    AggregateField? aggregateField2,
+    AggregateField? aggregateField3,
+    AggregateField? aggregateField4,
+    AggregateField? aggregateField5,
+    AggregateField? aggregateField6,
+    AggregateField? aggregateField7,
+    AggregateField? aggregateField8,
+    AggregateField? aggregateField9,
+    AggregateField? aggregateField10,
+    AggregateField? aggregateField11,
+    AggregateField? aggregateField12,
+    AggregateField? aggregateField13,
+    AggregateField? aggregateField14,
+    AggregateField? aggregateField15,
+    AggregateField? aggregateField16,
+    AggregateField? aggregateField17,
+    AggregateField? aggregateField18,
+    AggregateField? aggregateField19,
+    AggregateField? aggregateField20,
+    AggregateField? aggregateField21,
+    AggregateField? aggregateField22,
+    AggregateField? aggregateField23,
+    AggregateField? aggregateField24,
+    AggregateField? aggregateField25,
+    AggregateField? aggregateField26,
+    AggregateField? aggregateField27,
+    AggregateField? aggregateField28,
+    AggregateField? aggregateField29,
+    AggregateField? aggregateField30,
+  ]) {
+    final fields = [
+      aggregateField1,
+      aggregateField2,
+      aggregateField3,
+      aggregateField4,
+      aggregateField5,
+      aggregateField6,
+      aggregateField7,
+      aggregateField8,
+      aggregateField9,
+      aggregateField10,
+      aggregateField11,
+      aggregateField12,
+      aggregateField13,
+      aggregateField14,
+      aggregateField15,
+      aggregateField16,
+      aggregateField17,
+      aggregateField18,
+      aggregateField19,
+      aggregateField20,
+      aggregateField21,
+      aggregateField22,
+      aggregateField23,
+      aggregateField24,
+      aggregateField25,
+      aggregateField26,
+      aggregateField27,
+      aggregateField28,
+      aggregateField29,
+      aggregateField30,
+    ].whereType<AggregateField>();
+    return MethodChannelAggregateQuery(
+      this,
+      _pigeonParameters,
+      _pointer.path,
+      pigeonApp,
+      fields.map(
+        (e) {
+          if (e is query.count) {
+            return AggregateQuery(
+              type: AggregateType.count,
+            );
+          } else if (e is query.sum) {
+            return AggregateQuery(
+              type: AggregateType.sum,
+              field: e.field,
+            );
+          } else if (e is query.average) {
+            return AggregateQuery(
+              type: AggregateType.average,
+              field: e.field,
+            );
+          } else {
+            throw ArgumentError(
+                'Unsupported aggregate method ${e.runtimeType}');
+          }
+        },
+      ).toList(),
+      isCollectionGroupQuery,
+    );
+  }
+
+  // This method is not exposed in the public API, but can be used internally
+  @override
+  AggregateQueryPlatform sum(String field) {
+    return MethodChannelAggregateQuery(
+      this,
+      _pigeonParameters,
+      _pointer.path,
+      pigeonApp,
+      [
+        AggregateQuery(
+          type: AggregateType.sum,
+          field: field,
+        )
+      ],
+      isCollectionGroupQuery,
+    );
+  }
+
+  // This method is not exposed in the public API, but can be used internally
+  @override
+  AggregateQueryPlatform average(String field) {
+    return MethodChannelAggregateQuery(
+      this,
+      _pigeonParameters,
+      _pointer.path,
+      pigeonApp,
+      [
+        AggregateQuery(
+          type: AggregateType.average,
+          field: field,
+        )
+      ],
       isCollectionGroupQuery,
     );
   }