+++ packages/firebase_analytics/firebase_analytics_platform_interface/lib/src/analytics_event_item.dart
@@ -169,6 +169,6 @@ class AnalyticsEventItem {
 
   @override
   String toString() {
-    return '$AnalyticsEventItem($asMap)';
+    return '$AnalyticsEventItem(${asMap()})';
   }
 }