+++ packages/firebase_performance/firebase_performance/example/lib/main.dart
@@ -8,7 +8,6 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_performance/firebase_performance.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart';
-import 'package:pedantic/pedantic.dart';
 
 import 'firebase_options.dart';