+++ tests/integration_test/firebase_database/firebase_database_e2e_test.dart
@@ -4,7 +4,6 @@
 
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_database/firebase_database.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:tests/firebase_options.dart';
@@ -22,11 +21,8 @@ late FirebaseDatabase database;
 const emulatorPort = 9000;
 
 // Android device emulators consider localhost of the host machine as 10.0.2.2
-// so let's use that if running on Android.
-final emulatorHost =
-    (!kIsWeb && defaultTargetPlatform == TargetPlatform.android)
-        ? '10.0.2.2'
-        : 'localhost';
+// but should be automatically mapped by the useDatabaseEmulator function.
+const emulatorHost = 'localhost';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();