+++ packages/firebase_analytics/firebase_analytics_platform_interface/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 3.9.0
+
+ - **FEAT**(analytics,ios): add support for initiateOnDeviceConversionMeasurement using email address or phone number ([#11926](https://github.com/firebase/flutterfire/issues/11926)). ([9bc83f1e](https://github.com/firebase/flutterfire/commit/9bc83f1e6770d263092f6367ca3cfe46a83bb1b1))
+
+## 3.8.4
+
+ - Update a dependency to the latest release.
+
+## 3.8.3
+
+ - **FIX**(analytics): fix the toString method on AnalyticsEventItem to print out its content ([#11977](https://github.com/firebase/flutterfire/issues/11977)). ([5fe96ece](https://github.com/firebase/flutterfire/commit/5fe96ece22847e29d2007afe2f6b85b3d6ce4b94))
+
 ## 3.8.2
 
  - Update a dependency to the latest release.