+++ packages/firebase_in_app_messaging/firebase_in_app_messaging_platform_interface/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.4+16
+
+ - Update a dependency to the latest release.
+
+## 0.2.4+15
+
+ - Update a dependency to the latest release.
+
 ## 0.2.4+14
 
  - Update a dependency to the latest release.