+++ packages/firebase_storage/firebase_storage_platform_interface/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 5.1.3
+
+ - Update a dependency to the latest release.
+
+## 5.1.2
+
+ - Update a dependency to the latest release.
+
 ## 5.1.1
 
  - Update a dependency to the latest release.